pyproject.toml
[tool.poetry]
name = "iui-test"
version = "0.1.0"
description = ""
authors = ["hoon <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
selenium = "^4.27.1"
webdriver-manager = "^4.0.2"
pillow = "^11.0.0"
graphviz = "^0.20.3"
retrying = "^1.3.4"

# 연준님
aiohappyeyeballs = "2.4.4"
aiohttp = "3.11.8"
aiosignal = "1.3.1"
aisuite = "0.1.6"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
appnope = "0.1.4"
asttokens = "2.4.1"
attrs = "24.2.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
comm = "0.2.2"
datasets = "3.1.0"
debugpy = "1.6.7"
decorator = "5.1.1"
dill = "0.3.8"
distro = "1.9.0"
exceptiongroup = "1.2.2"
executing = "2.1.0"
filelock = "3.16.1"
frozenlist = "1.5.0"
fsspec = "2024.9.0"
h11 = "0.14.0"
httpcore = "1.0.7"
httpx = "0.28.0"
huggingface-hub = "0.26.3"
idna = "3.10"
importlib_metadata = "8.5.0"
ipykernel = "6.29.5"
ipython = "8.30.0"
jedi = "0.19.2"
jiter = "0.8.0"
jupyter_client = "8.6.3"
jupyter_core = "5.7.2"
Markdown = "3.7"
matplotlib-inline = "0.1.7"
multidict = "6.1.0"
multiprocess = "0.70.16"
nest_asyncio = "1.6.0"
numpy = "2.1.3"
openai = "1.55.3"
packaging = "24.2"
pandas = "2.2.3"
parso = "0.8.4"
pexpect = "4.9.0"
pickleshare = "0.7.5"
pip = "24.2"
platformdirs = "4.3.6"
prompt_toolkit = "3.0.48"
propcache = "0.2.0"
psutil = "6.1.0"
ptyprocess = "0.7.0"
pure_eval = "0.2.3"
pyarrow = "18.1.0"
pydantic = "2.10.2"
pydantic_core = "2.27.1"
Pygments = "2.18.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.2"
PyYAML = "6.0.2"
pyzmq = "25.1.2"
requests = "2.32.3"
setuptools = "75.1.0"
six = "1.16.0"
sniffio = "1.3.1"
stack-data = "0.6.2"
tornado = "6.4.2"
tqdm = "4.67.1"
traitlets = "5.14.3"
typing_extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.3"
wcwidth = "0.2.13"
wheel = "0.44.0"
xxhash = "3.5.0"
yarl = "1.18.0"
zipp = "3.21.0"
importlib = "^1.0.4"
html5lib = "^1.1"
breadability = {git = "https://github.com/bookieio/breadability.git"}
summy = "^0.1"
konlpy = "^0.6.0"
sumy = "^0.11.0"
scipy = "^1.14.1"
nltk = "^3.9.1"
mercury-parserpy = "^0.7.0"
fake-useragent = "^2.0.0"
readability = "^0.3.1"
readability-lxml = "^0.8.1"
transformers = "^4.49.0"
torch = "^2.6.0"
protobuf = "^5.29.3"
[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
